allow autoregression in optim hfc for torch models
