+++ darts/utils/historical_forecasts/__init__.py
@@ -3,6 +3,7 @@ from .optimized_historical_forecasts_regression import (
     _optimized_historical_forecasts_last_points_only,
 )
 from .utils import (
+    _check_optimizable_historical_forecasts_global_models,
     _get_historical_forecast_boundaries,
     _historical_forecasts_general_checks,
     _historical_forecasts_start_warnings,
